+++ .idea/WindowsBackgrounds.iml
@@ -2,13 +2,18 @@
 <module type="PYTHON_MODULE" version="4">
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/Project/Testing" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/Testing" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/0 - Bin/venv" />
       <excludeFolder url="file://$MODULE_DIR$/build" />
-      <excludeFolder url="file://$MODULE_DIR$/Project/Testing/Backgrounds" />
       <excludeFolder url="file://$MODULE_DIR$/build/dist" />
+      <excludeFolder url="file://$MODULE_DIR$/Testing/Backgrounds" />
+      <excludeFolder url="file://$MODULE_DIR$/.venv" />
     </content>
-    <orderEntry type="inheritedJdk" />
+    <orderEntry type="jdk" jdkName="Python 3.11 (WindowsBackgrounds)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+  <component name="PackageRequirementsSettings">
+    <option name="removeUnused" value="true" />
   </component>
 </module>
