--- Project/bin/Resources/Settings_template.cfg
@@ -1,2 +1,0 @@
-[VARIABLES]
-UseWindowsBackgrounds = False
