+++ src/wxglade/app.py.bak
@@ -1,7 +1,7 @@
-#!/usr/bin/env python3
+#!/usr/src/env python3
 # -*- coding: UTF-8 -*-
 #
-# generated by wxGlade 1.0.4 on Fri Nov  3 14:43:28 2023
+# generated by wxGlade 1.0.4 on Thu Nov  2 15:52:24 2023
 #
 
 # This is an automatically generated file.
