+++ src/wxglade/wxFrameFolderPaths.py.bak
