+++ src/wxglade/wxFrameMainApp.py
