+++ src/wxglade/wxFrameMainApp.py.bak
