+++ src/wxglade/wxMyFrame.py.bak
