+++ src/wxglade/wxRenameDialog.py
