+++ src/wxglade/wxRenameDialog.py.bak
