+++ src/wxglade/wxTabTemplate.py
