+++ src/wxglade/wxTabTemplate.py.bak
