+++ src/wxglade/wxWarningDialog.py
