+++ src/wxglade/wxWarningDialog.py.bak
