+++ src/wxglade_overrides/RenameDialog.py
