+++ src/wxglade_overrides/WarningDialog.py
